feat: added unit tests for board generation method
